backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic-settings
email-validator
PyJWT==2.8.0
python-multipart==0.0.6
passlib[bcrypt]
pytest
httpx
